requirements.txt
annotated-types==0.6.0
anyio==4.3.0
fastapi==0.110.1
idna==3.6
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python_dotenv==1.0.1
python_multipart==0.0.9
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0